seperate ses arn role from create email address method
