+++ src/utils/notifications.py
@@ -1,10 +1,18 @@
 """Email Notifications."""
 # Third-Party Libraries
+import boto3
+from botocore.exceptions import ClientError
 from flask import g, render_template
 
 # cisagov Libraries
 from api.manager import UserManager
-from utils.aws.ses import send_message
+from settings import APP_ENV, APP_NAME, SES_ASSUME_ROLE_ARN, SMTP_FROM
+from utils.aws import sts
+
+if SES_ASSUME_ROLE_ARN:
+    ses = sts.assume_role_client("ses", SES_ASSUME_ROLE_ARN)
+else:
+    ses = boto3.client("ses")
 
 user_manager = UserManager()
 
@@ -51,3 +59,32 @@ class Notification:
             text=content["text_content"],
             html=content["html_content"],
         )
+
+
+def send_message(to: str, subject: str, text: str, html: str):
+    """Send message via SES."""
+    resp = {}
+    try:
+        resp = ses.send_email(
+            Source=SMTP_FROM,
+            Destination={
+                "ToAddresses": [
+                    to,
+                ],
+            },
+            Message={
+                "Subject": {"Data": subject, "Charset": "UTF-8"},
+                "Body": {
+                    "Text": {"Data": text, "Charset": "UTF-8"},
+                    "Html": {"Data": html, "Charset": "UTF-8"},
+                },
+            },
+            Tags=[
+                {"Name": "app", "Value": APP_NAME},
+                {"Name": "environment", "Value": APP_ENV},
+            ],
+        )
+    except ClientError as e:
+        return e.response["Error"]
+
+    return resp
